Various fixes
First and foremost: Get it to work.
Update packages.
Give floating_view a background to see where to drag to move it.
Add a End button to quit using the autoclicker service.
Change onCreate to onStart so it checks accessibility and draw on top permissions each time you go back to the main app.
Include a signed prebuilt apk for android 12.